question_visual_agent/requirements.txt
openai>=1.104.1,<2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.31.0